+++ src/main/java/me/thegiggitybyte/sleepwarp/mixin/common/ServerWorldMixin.java
@@ -1,4 +1,4 @@
-package me.thegiggitybyte.sleepwarp.mixin;
+package me.thegiggitybyte.sleepwarp.mixin.common;
 
 import me.thegiggitybyte.sleepwarp.SleepWarp;
 import net.minecraft.entity.player.PlayerEntity;
@@ -10,6 +10,7 @@ import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.server.world.ServerWorld;
 import net.minecraft.server.world.SleepManager;
 import net.minecraft.util.profiler.Profiler;
+import net.minecraft.village.raid.RaidManager;
 import net.minecraft.world.GameRules;
 import net.minecraft.world.MutableWorldProperties;
 import net.minecraft.world.World;
@@ -24,7 +25,6 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.Redirect;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.function.BooleanSupplier;
 import java.util.function.Supplier;
@@ -42,6 +42,8 @@ public abstract class ServerWorldMixin extends World {
     @Shadow @NotNull public abstract MinecraftServer getServer();
     @Shadow protected abstract void wakeSleepingPlayers();
     @Shadow protected abstract void resetWeather();
+    
+    @Shadow public abstract RaidManager getRaidManager();
     
     @Inject(method = "tick", at = @At(value = "INVOKE_ASSIGN", target = "Lnet/minecraft/world/GameRules;getInt(Lnet/minecraft/world/GameRules$Key;)I"))
     private void trySleepWarp(BooleanSupplier shouldKeepTicking, CallbackInfo ci) {
