+++ src/main/resources/sleepwarp.mixins.json
@@ -4,9 +4,13 @@
   "compatibilityLevel": "JAVA_17",
   "package": "me.thegiggitybyte.sleepwarp.mixin",
 
+  "client": [
+
+  ],
+
   "mixins": [
-    "ServerWorldMixin",
-    "SleepManagerAccessor"
+    "common.ServerWorldMixin",
+    "common.SleepManagerAccessor"
   ],
 
   "injectors": {
